Update tests and clean up: add new test cases, modify existing ones, and remove obsolete files

+++ MiniC/TypingAndInterpret/tests/students/control_flow/conditional/if_else_statement.c
@@ -1,7 +1,7 @@
 #include "printlib.h"
 
 int main() {
-    float x;
+    int x;
     x = 9;
 
     if(x > 10) {
